# backend/openapi.yaml
openapi: 3.1.0
info:
  title: GYM Training API
  version: 0.2.0
  description: Authenticated sync API for the GYM local-first training product.
servers:
  - url: https://api.example.invalid
paths:
  /v1/health:
    get:
      operationId: health
      responses:
        "200":
          description: Service health
  /v1/auth/register:
    post:
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 12, maxLength: 128 }
      responses:
        "201": { description: Account created }
        "409": { description: Email already exists }
  /v1/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 12, maxLength: 128 }
      responses:
        "200": { description: Authenticated }
        "401": { description: Invalid credentials }
  /v1/me:
    get:
      security: [{ bearerAuth: [] }]
      operationId: me
      responses:
        "200": { description: Current athlete }
        "401": { description: Authentication required }
  /v1/sessions:
    post:
      security: [{ bearerAuth: [] }]
      operationId: createSession
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, minLength: 16, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SessionCreate" }
      responses:
        "201": { description: Session persisted }
        "200": { description: Idempotent replay }
        "400": { description: Invalid request }
        "409": { description: Conflicting resource }
    get:
      security: [{ bearerAuth: [] }]
      operationId: listSessions
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200": { description: Session page }
  /v1/movement-events:
    post:
      security: [{ bearerAuth: [] }]
      operationId: createMovementEvent
      description: Store validated derived movement metrics. Raw camera video is outside this API by default.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MovementEvent" }
      responses:
        "201": { description: Event accepted }
        "404": { description: User does not own the session }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    SessionCreate:
      type: object
      required: [id, startedAt, source, exercises]
      properties:
        id: { type: string }
        startedAt: { type: string, format: date-time }
        source: { type: string, enum: [manual, camera, wearable] }
        name: { type: string, maxLength: 120 }
        exercises:
          type: array
          maxItems: 50
          items: { $ref: "#/components/schemas/ExerciseSession" }
    ExerciseSession:
      type: object
      required: [exerciseId, sets]
      properties:
        exerciseId: { type: string }
        sets:
          type: array
          maxItems: 50
          items: { $ref: "#/components/schemas/WorkoutSet" }
    WorkoutSet:
      type: object
      required: [index, reps, weightKg, completed]
      properties:
        index: { type: integer, minimum: 1 }
        reps: { type: number, minimum: 0, maximum: 1000 }
        weightKg: { type: number, minimum: 0, maximum: 1000 }
        completed: { type: boolean }
        completedAt: { type: string, format: date-time, nullable: true }
        rpe: { type: number, minimum: 1, maximum: 10, nullable: true }
    MovementEvent:
      type: object
      required: [schemaVersion, sessionId, exerciseId, timestamp, source, reps]
      properties:
        schemaVersion: { type: integer, minimum: 1 }
        sessionId: { type: string }
        exerciseId: { type: string }
        timestamp: { type: string, format: date-time }
        source: { type: string, enum: [manual, camera, wearable] }
        reps: { type: number, minimum: 0, maximum: 1000 }
        confidence: { type: number, minimum: 0, maximum: 1, nullable: true }
        model: { type: string, nullable: true }
        metrics: { type: object, additionalProperties: true }